move to agentmetrics class
